% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/4010Project1.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle Function}
\usage{
shuffle(cards)
}
\arguments{
\item{cards}{vector or array}
}
\value{
cards in a random order
}
\description{
Shuffle Function
}
\examples{
site <- "https://gist.githubusercontent.com/garrettgman/9629323/raw/ee5dfc039fd581cb467cc69c226ea2524913c3d8/deck.csv"
cards <- readr::read_csv(site)
shuffle(cards)
}

